+++ fengyu-staff/cloudfunctions/staffApi/utils/points.js
@@ -65,7 +65,7 @@ async function settlePointsForOrder(client, originalSaleOrderId) {
 
   // 4. 已发积分合计（按 ref_order_id 聚合，等级升级/兑换等非本链流水自然排除）
   const grantedRes = await client.query(
-    `SELECT COALESCE(SUM(amount), 0)::int AS granted
+    `SELECT COALESCE(SUM(amount), 0)::bigint AS granted
        FROM point_transactions
       WHERE ref_order_id = $1`,
     [originalSaleOrderId],
